requirements.txt
python-telegram-bot==20.3
apscheduler
flask
openai>=1.40.0